set layout and seprate  routes
